file: support handle large file

+++ src/connection.c
@@ -62,7 +62,7 @@ static void conn_send(struct uh_connection *conn, const void *data, ssize_t len)
     ev_io_start(conni->srv->loop, &conni->iow);
 }
 
-static void conn_send_file(struct uh_connection *conn, const char *path, size_t offset, ssize_t len)
+static void conn_send_file(struct uh_connection *conn, const char *path, off_t offset, int64_t len)
 {
     struct uh_connection_internal *conni = (struct uh_connection_internal *)conn;
     size_t min = 8192;
