+++ src/uhttpd.h
@@ -57,7 +57,7 @@ struct uh_connection {
     */
     void (*done)(struct uh_connection *conn);
     void (*send)(struct uh_connection *conn, const void *data, ssize_t len);
-    void (*send_file)(struct uh_connection *conn, const char *path, size_t offset, ssize_t len);
+    void (*send_file)(struct uh_connection *conn, const char *path, off_t offset, int64_t len);
     void (*printf)(struct uh_connection *conn, const char *format, ...) __attribute__((format(printf, 2, 3)));
     void (*vprintf)(struct uh_connection *conn, const char *format, va_list arg);
     void (*send_status_line)(struct uh_connection *conn, int code, const char *extra_headers);
